requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
PyJWT==2.9.0
Werkzeug==3.0.3
